requirements.txt
requests==2.32.3
redis[asyncio]
httpx==0.27.2
pyyaml
fastapi==0.115.4
uvicorn[standard]
fastapi-limiter
pydantic
loguru==0.7.2
python-dotenv
json-repair
aiofiles
email-validator